refactor(core): simplify keypair transaction scaffolding
