+++ .changeset/drain-scoped-transaction-work.md
@@ -10,7 +10,3 @@ independent operations can use `Promise.all()` without manually draining sibling
 Preserve repository bindings exposed through inherited properties or class getters when binding a
 transaction scope, including the original receiver for getters. Frozen repositories with own
 methods also retain lifetime protection without violating JavaScript proxy invariants.
-
-Remove the scoped allocation queue. Within one transaction, implementations await dependent
-mutations sequentially and only parallelize independent work. Concurrent standalone allocations
-remain isolated by their separate transactions.
