+++ packages/core/test/unit/KeyRingService.test.ts
@@ -16,6 +16,7 @@ import { RepositoryCoreTransactionRunner } from '../../transactions/CoreTransact
 import { CoreKeyRingTransactions } from '../../transactions/keypairs/KeyRingTransactions.ts';
 import { KeypairDerivation } from '../../keypairs/KeypairDerivation.ts';
 import { KeypairP2pkSigner } from '../../keypairs/P2pkSigner.ts';
+import { overrideTransactions } from '../overrideTransactions.ts';
 
 // Mock seed for deterministic testing
 const MOCK_SEED = new Uint8Array(64);
@@ -39,19 +40,6 @@ describe('KeyRingService', () => {
       new KeypairDerivation(() => seed.getSeed()),
       new KeypairP2pkSigner(transactionRepositories.keyRingRepository),
     );
-  }
-
-  function overrideTransactions(
-    base: Repositories,
-    withTransaction: Repositories['withTransaction'],
-  ): Repositories {
-    return new Proxy(base, {
-      get(target, property, receiver) {
-        if (property === 'withTransaction') return withTransaction;
-        const value = Reflect.get(target, property, receiver) as unknown;
-        return typeof value === 'function' ? value.bind(target) : value;
-      },
-    });
   }
 
   beforeEach(() => {
